requirements.txt
# Database Drivers
oracledb>=2.0.0
mysql-connector-python>=8.0.0

# NATS Client (synchronous with JetStream support)
nats-py>=2.6.0

# Configuration
PyYAML>=6.0.1
